Deploy LLM-CostOps to Google Cloud Run
- Add serve command to CLI for HTTP server mode
- Add agent API routes (budget-enforcement, cost-forecasting, etc.)
- Add Dockerfile.cloudrun optimized for Cloud Run
- Add Cloud Build configurations (dev, staging, prod)
- Add GCP deployment scripts and IAM setup
- Fix rmp-serde edition2024 compatibility (use rust:latest)
- Fix async recursion in ruvector client with Box::pin
- Add BudgetConfig export to forecasting module
- Add SerializationError variant to BudgetEnforcementError
- Fix Redis expire type (i64 -> usize)
- Update .gitignore to exclude Claude files and GCP SDK
- Update .gcloudignore to not exclude source metrics directory

Service deployed to: https://llm-costops-dev-xx7kwyd5ra-uc.a.run.app

+++ crates/llm-cost-ops-api/src/ingestion/ratelimit.rs
@@ -450,7 +450,7 @@ impl RateLimiter for RedisRateLimiter {
                 .map_err(|e| CostOpsError::Integration(format!("Redis zadd failed: {}", e)))?;
 
             // Set expiry on the key
-            let expiry = (config.window_duration.as_secs() + 60) as i64;
+            let expiry = (config.window_duration.as_secs() + 60) as usize;
             let _: () = con.expire(&key, expiry).await
                 .map_err(|e| CostOpsError::Integration(format!("Redis expire failed: {}", e)))?;
